+++ commons/src/lib.rs
@@ -1,15 +1,13 @@
 #![warn(unused_crate_dependencies)]
 #![warn(clippy::pedantic)]
 
+pub mod build_output;
 pub mod cache;
 pub mod display;
 pub mod fun_run;
-pub mod gem_list;
 pub mod gem_version;
 pub mod gemfile_lock;
 pub mod layer;
 pub mod metadata_digest;
-pub mod rake_status;
-pub mod rake_task_detect;
 
 mod err;